Factor out postgres service and fix postgres
